refactor: move UserIconCollectionDB interface
